server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.2
textblob==0.17.1
# ML and Deep Learning
transformers==4.36.2
torch==2.1.2
tensorflow==2.15.0
sentence-transformers==2.2.2
bertopic==0.15.0
scikit-learn==1.3.2
umap-learn==0.5.5
hdbscan==0.8.33
# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
# Utilities
numpy==1.24.3
tqdm==4.66.1
joblib==1.3.2 